test_training/hand_written_digits/requirements.txt
flask
flask-cors
tensorflow
pillow
opencv-python-headless
numpy
